requirements.txt
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
boto3==1.9.185
botocore==1.12.185
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.3
django-cors-headers==3.0.2
django-debug-toolbar==2.0
django-extensions==2.1.9
django-rest-swagger==2.1.2
django-storages==1.7.1
djangorestframework==3.9.4
docutils==0.14
drf-yasg==1.16.1
entrypoints==0.3
flex==6.14.0
iamport-rest-client==0.7.0
idna==2.8
inflection==0.3.1
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
itypes==1.1.0
jedi==0.14.1
Jinja2==2.10.1
jmespath==0.9.4
jsonfield==2.0.2
jsonpointer==2.0
jsonschema==3.0.1
jupyter-client==5.3.1
jupyter-core==4.5.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
openapi-codec==1.3.2
packaging==19.0
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psycopg2-binary==2.8.3
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.0
pyrsistent==0.15.3
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
pyzmq==18.0.2
requests==2.22.0
rfc3987==1.3.8
ruamel.yaml==0.15.100
s3transfer==0.2.1
Send2Trash==1.5.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
strict-rfc3339==0.7
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.3
validate-email==1.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.0
